+++ file-sharing/src/composables/useUserGroupLists.js
@@ -1,98 +1,78 @@
-import { ref, onBeforeMount, onUnmounted } from 'vue';
+import { ref, onBeforeMount, onUnmounted, watch } from 'vue';
+import { useConfig } from '../components/Config.vue';
 import { useSpawn, errorString } from '@45drives/cockpit-helpers';
 
 const spawnOpts = {
 	superuser: 'try',
-}
-
-/**
- * @param {String} type - 'group' or 'user'
- * @param {boolean} domain 
- * @returns 
- */
-const parseRecord = (type, domain) => (record) => {
-	if (!record)
-		return null;
-	let nameKey = type;
-	let idKey = type === 'group' ? 'gid' : 'uid';
-
-	const fields = record.split(':');
-	const name = fields[0];
-	const id = fields[2];
-	if (id >= 1000 || id === '0') // include root
-		return { [nameKey]: name, [idKey]: id, domain, pretty: name + (domain ? ' (domain)' : '') };
-	return null;
 };
 
-const getent = async (db) => {
-	const [name, wbinfoFlag] = db === 'passwd' ? ['user', '-u'] : ['group', '-g'];
-	const procs = [];
-	procs.push(
-		useSpawn(['getent', '-s', 'files', db], spawnOpts).promise()
-			.then(({ stdout: getentDb }) =>
-				getentDb.trim()
-					.split('\n')
-					.map(parseRecord(name, false))
-			)
-	);
-	await useSpawn(['wbinfo', wbinfoFlag], spawnOpts).promise()
-		.then(({ stdout }) => {
-			const domainItemNames = stdout.trim().split('\n');
-			// limit number of queries per execution of getent to avoid
-			// sending too much data and causing cockpit ws to drop connection
-			const partitionSize = 1000;
-			for (let i = 0; i < domainItemNames.length; i += partitionSize) {
-				procs.push(
-					useSpawn(['getent', db, ...domainItemNames.slice(i, i + partitionSize)], spawnOpts).promise()
-						.catch(state => {
-							console.error('getent error while getting domain entries:', state);
-							return state;
-						})
-						.then(({ stdout: getentDb }) =>
-							getentDb.trim()
-								.split('\n')
-								.map(parseRecord(name, true))
-						)
-				);
-			}
-		})
-		.catch(e => console.warn(`Error getting list of ${name}s from domain. wbinfo:`, errorString(e)));
-	return await Promise.all(procs)
-		.then(nestedEntries =>
-			nestedEntries.flat().filter(entry => entry !== null).sort((a, b) => a.pretty.localeCompare(b.pretty))
-		);
-}
+/**
+ * Get bash script for json ouput of users or groups
+ * 
+ * @param {"passwd" | "group"} db - getent database to query
+ * @param {boolean} isDomain - if true, pull from domain users, else local users
+ * @param {boolean} includeSystemAccounts - keep users/groups with uid/gid between 0 and 1000
+ * @returns {string} bash script
+ */
+const getentScript = (db, isDomain, includeSystemAccounts) => {
+	const [nameKey, idKey, wbinfoFlag] = {
+		'passwd': ['user', 'uid', '-u'],
+		'group': ['group', 'gid', '-g']
+	}[db];
+	const getent = isDomain ? `wbinfo ${wbinfoFlag} | xargs -d'\n' -r getent` : `getent -s files`;
+	const printFilter = includeSystemAccounts ? '' : '$3 >= 1000 || $3 == 0';
+	return `set -o pipefail; ${getent} ${db} | awk -F: '
+		BEGIN { printf "[" }
+		${printFilter} {
+			printf sep;
+			printf "{";
+			printf "\\"${nameKey}\\":\\"%s\\",",$1;
+			printf "\\"${idKey}\\":%d,",$3;
+			printf "\\"domain\\":${isDomain.toString()},";
+			printf "\\"pretty\\":\\"%s${isDomain ? ' (domain)' : ''}\\"",$1;
+			printf "}";
+			sep = ",";
+		}
+		END { printf "]" }
+		'`;
+};
+
+const getent = (db, includeSystemAccounts) => {
+	return Promise.all([
+		useSpawn(['bash', '-c', getentScript(db, false, includeSystemAccounts)], spawnOpts).promise()
+			.then(({ stdout }) => {
+				return JSON.parse(stdout);
+			}),
+		useSpawn(['bash', '-c', getentScript(db, true, includeSystemAccounts)], spawnOpts).promise()
+			.catch(state => {
+				console.warn(`Error getting users/groups from AD/domain:`, errorString(state));
+				return state;
+			})
+			.then(({ stdout }) => JSON.parse(stdout)),
+	]).then(nested => nested.flat());
+};
 
 /**
- * get list of user objects from system and domain if joined
- * 
- * @returns {Promise<Object[]>}
+ * @typedef User
+ * @property {string} name - system name of user
+ * @property {number} uid - user id
+ * @property {boolean} domain - if true, user is from AD
+ * @property {string} pretty - 
  */
-function getUsers() {
-	return getent('passwd');
-}
-
-/**
- * get list of group objects from system and domain if joined
- * 
- * @returns {Promise<Object[]>}
- */
-function getGroups() {
-	return getent('group');
-}
 
 export default function useUserGroupLists() {
 	const users = ref([]);
 	const groups = ref([]);
 	const processingUsersList = ref(false);
 	const processingGroupsList = ref(false);
+	const config = useConfig();
 
 	const cockpitWatchHandles = [];
 
 	async function aquireUserList() {
 		processingUsersList.value = true;
 		try {
-			users.value = await getUsers();
+			users.value = await getent('passwd', config.includeSystemAccounts ?? false);
 		} catch (state) {
 			console.error('Failed to get user list:', errorString(state));
 		} finally {
@@ -103,7 +83,7 @@ export default function useUserGroupLists() {
 	async function aquireGroupList() {
 		processingGroupsList.value = true;
 		try {
-			groups.value = await getGroups();
+			groups.value = await getent('group', config.includeSystemAccounts ?? false);
 		} catch (state) {
 			console.error('Failed to get group list:', errorString(state));
 		} finally {
@@ -116,11 +96,15 @@ export default function useUserGroupLists() {
 		cockpitWatchHandles.push(
 			cockpit.file('/etc/passwd', { superuser: 'try' }).watch(() => aquireUserList(), { read: false }),
 			cockpit.file('/etc/group', { superuser: 'try' }).watch(() => aquireGroupList(), { read: false })
-		)
+		);
+		watch(() => config.includeSystemAccounts, () => {
+			aquireUserList();
+			aquireGroupList();
+		});
 	});
 
 	onUnmounted(() => {
-		cockpitWatchHandles.forEach(handle => handle?.remove?.())
+		cockpitWatchHandles.forEach(handle => handle?.remove?.());
 	});
 
 	return {
@@ -128,5 +112,5 @@ export default function useUserGroupLists() {
 		groups,
 		processingUsersList,
 		processingGroupsList,
-	}
+	};
 }